AutoCloudSkill.exe_extracted/requirements.txt
# Core Dependencies
playwright==1.47.0
google-genai>=0.6.0
google-api-python-client>=2.183.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0

# Web & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
certifi>=2023.0.0
idna>=3.4
charset-normalizer>=3.3.0

# GUI
ttkbootstrap>=1.10.0

# Audio Processing
speechrecognition>=3.10.0
imageio-ffmpeg>=0.4.9
moviepy>=1.0.3

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0

# System
machineid>=0.3.0

# Optional (for development)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
